Clea code: Fix spelling error "interactive launcher".
