+++ docs/sandbox/InteractiveOtpMain.md
@@ -1,27 +1,5 @@
 # Interactive OTP Launcher
-A GUI popup window to which help you to start OTP Main in an interactive way.
-The UI look "like" this:
-```
-Select OTP data source
-  ( ) dev-2.x/osm-no
-  (x) dev-2.x/entur
-  ( ) dev-2.x/sf-bay-area
-  ( ) 2.0/osm-no
-  ( ) 2.0/entur
-  ( ) 2.0/sf-bay-area
-
-Build
-  [ ] Street Graph
-  [ ] Transit Graph
-
-Actions
-  [ ] Save
-  [ ] Serve
-
-
-[ Start OTP ]
-
-```
+A GUI popup window to which help you to start OTP Main interactively.
 
 
 ## Contact Info
