+++ src/ext/java/org/opentripplanner/ext/interactivelauncher/SetupResult.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.ext.interactivlancher;
+package org.opentripplanner.ext.interactivelauncher;
 
 import java.io.File;
 import java.util.ArrayList;
