+++ src/ext/java/org/opentripplanner/ext/interactivelauncher/views/StartOtpButtonView.java
@@ -1,10 +1,10 @@
-package org.opentripplanner.ext.interactivlancher.views;
+package org.opentripplanner.ext.interactivelauncher.views;
 
 import javax.swing.*;
 
 import java.awt.event.ActionListener;
 
-import static org.opentripplanner.ext.interactivlancher.views.ViewUtils.adjustSize;
+import static org.opentripplanner.ext.interactivelauncher.views.ViewUtils.adjustSize;
 
 class StartOtpButtonView {
   private static final int BUTTON_D_WIDTH = 160;
