+++ src/ext/java/org/opentripplanner/ext/interactivelauncher/views/ViewUtils.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.ext.interactivlancher.views;
+package org.opentripplanner.ext.interactivelauncher.views;
 
 import javax.swing.*;
 import java.awt.*;
